[BUGFIX] Drop a strange internal DataHandler method
That's a seldom gem.

Method convNumEntityToByteValue() in DataHandler
is *seriously* strange. It exists since very early
TYPO3 claiming to fix a "very rare" MSIE bug. It
*tries* to convert the "&#123;" of a string like
"foo&#123;bar" to an ascii char ("{") and does
this with a weird logic that includes a microtime().

Testing this thing locally reveals the method
returns different results per execution!

The patch drops that @internal method and it's
caller.

Change-Id: If9da47cc39d7c811b80879a4b532af90e9ea6f20
Resolves: #95789
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71985
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>