[BUGFIX] Sanitize unserialize() in ProcessedFile
Table sys_file_processedfile field configuration is a
blob. It usually contains a serialized array with
processing information, but can be null due to the
nature of blob. unserialize(null) emits an E_DEPRECATED
error with PHP 8.1. ProcessedFileTest stumbles here,
we however don't fix that in the test, but in the code
to deal with the fact that this table field might be
null in DB for whatever reasons.

Resolves: #95790
Releases: master
Change-Id: I961f042088347ade2689fb2c085543373d1104f2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71986
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>