[BUGFIX] Fix Frontend translations for file metadata in free mode
When rendering a frontend in a language set to free
mode (= just do a SQL query for e.g. language=5), sys_file_metadata
is now also shown in the frontend's language.

The logic within FileMetadataAspect is always loading the
default language (L=0) and for this, we need to call
PageRepository->getLanguageOverlay() for "sys_file_metadata"
and even when running in free mode, handle a "translation"
for sys_file_metadata just like we'd do overlays.

This is the same concept that TYPO3 needs to use for pages
(page always resolves to a default language record).

Resolves: #93025
Releases: main, 13.4
Change-Id: If0ee2e304dbbd1a4fa0a30ee6f17c414d98cb54e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72015
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>