[BUGFIX] Do not show empty colPos message for special unused column
Columns, not having a colPos defined, are considered as
inaccessible. A event listener therefore enriches such
columns with a corresponding information in the page
module. Since the special "unused elements" column is
obviously also missing the "colPos", the information
was previously also shown for this column.

This is now fixed by checking for the special "unused"
column, before enriching the columns markup.

Resolves: #95870
Releases: master
Change-Id: I7af75e58181af24d772aa63a3fccec5c2d52c5e8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72050
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>