[BUGFIX] Fix undefined key warning in EXT:frontend with PHP8
When using a text content element with a table inside bodytext,
you will have a PHP warning: Undefined array key "1." in
typo3/sysext/frontend/Classes/ContentObject/ContentObjectRenderer.php

This was only a notice with PHP v7 but it's an error with PHP v8.
This bugfix corrects the usage of undefined array keys.

Resolves: #95858
Releases: master
Change-Id: Ie70fb6bdd66eee895e0c290a9bcc662c9f734d2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72058
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>