[TASK] Move Composer package artifact to vendor folder
Storing the package artifact in the var/build folder raised
questions how this folder shall be treated in general and during
deployment. Since the Package Artifact is built in
a `composer install` process, it should be treated the same as
other build artifacts like all files in the vendor folder.

To make this more clear, the package artifact is now also stored in
the vendor folder.

Accessing the vendor folder during runtime is now possible with
using Composer 2.1 runtime API, which is required anyway already.

The path to the vendor folder is determined indirectly, by using
the path to the typo3/cms-composer-installers package. This means,
this package must not be replaced by another package with a different
name. Since no public replacements for this package exist and
there are no benefits for a private replacement of this package over
having a private fork with the same package name,
this should not be an impediment.

To ease the required related changes in the testing framework,
late static binding is used for calling
SystemEnvironmentBuilder::usesComposerClassLoading

Resolves: #95897
Releases: master
Change-Id: I3cfc982c289b312d22f59d2704d7e7932efc8cc1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72071
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>