[BUGFIX] Avoid InvalidTypoScriptConditionException in RedirectHandler
Redirects to target with LinkHandler uri 't3://' need a bootstrapped
'TypoScriptFrontendController' and parsed and executed TypoScript,
which is done later in the middleware stack for normal requests. Thus
RedirectService bootstraps 'TypoScriptFrontedController' to generate
uri for LinkHandler uri.

At this point global '$GLOBAL['TYPO3_REQUEST']' has not been in set,
and the fallback to create it from GLOBALS would not have correct
site and siteLanguage attributes set, which would fail to evaluate
TypoScript conditions correctly like 'fallbackLanguageIds' array
check for siteLanguage attribute, as null is returned instead of
an array.

For v11 this was already fixed through moving to passing the PSR-7
request and initalizing TSFE with passed request in #92666 and the
coresponding patch.

This patch sets a '$GLOBAL['TYPO3_REQUEST'] to the current '$request'
to avoid 'InvalidTypoScriptConditionException' if page is in FE debug
context (ex. DebugPreset). If no redirect is found to be handled, that
temporarly global request object is removed, to handle it through the
normal middleware stack.

Resolves: #95882
Related: #92666
Releases: 10.4
Change-Id: I6d7bff643fa2440e3986e4362c1882d53db1bb67
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72074
Tested-by: core-ci <typo3@b13.com>
Tested-by: Sybille Peters <sypets@gmx.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>