[BUGFIX] Avoid call mb_strlen(null) in GeneralUtility::fixed_lgd_cs
GeneralUtility::fixed_lgd_cs() may be called with null instead of a
string as first argument, which is passed to mb_strlen(). Feeding
null to mb_strlen() emits a E_DEPRECATED PHP 8.1 error.

This patch sanitizes the call in BackendUtillity and further ensures
that the first parameter $string is a string (until strict type hints
can be added).

Resolves: #95913
Releases: master
Change-Id: Ie3bb44d4d47d4673d2274f10dca80f860d494c32
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72085
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Sybille Peters <sypets@gmx.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>