[BUGFIX] Replace non existing icon in DB Check
It's possible to query for deleted records in DB Check
"Full search". Such records can then be restored only
or additionally restored and directly edited. The icon
for the latter action was broken, since a non existing
icon identifier was used.

This is now fixed, using the correct icon identifier.

Resolves: #95923
Releases: master, 10.4
Change-Id: I7da25198efe731a9ccba9e673c8c55e42760b233
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72100
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>