[BUGFIX] Make toolbar dropdown menu scrollable
The contents of the toolbar dropdown menus
for bookmarks and open documents are dynamic.
Having e.g. a lot bookmarks previously lead to
the most recent bookmarks might be cut of at
the bottom of the screen.

This is now fixed by using a maximum height
and allowing vertical scroll.

Resolves: #94641
Releases: master
Change-Id: I50b799c628e51fbcb8b59ad5eb64d369647c27fe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72140
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>