[TASK] Add correct test for shortcut issue in language menu
The added tests for #89871 do not prove the
correctness of the patch. They go green, without the
fix being applied. This is a blocker for another
regression bugfix, which depends on these tests.

The issue lied in the language menu, not the normal
menu. For this, a new method createLanguageMenu has
been added to AbstractLocalizedPagesTestCase.php. It
is now used to assure the generated links always
point to the target language and not to the current
one.

Resolves: #95966
Related: #89871
Related: #94677
Releases: master
Change-Id: I20d202c2eae58597ee59312ef0896debb2f93d39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72162
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>