[BUGFIX] Always display root page in page tree
This applies to rendering of database mounts (from groups or
from workspaces): They are now shown as children of the virtual
root page (uid=0) as it used to be in TYPO3 v8.

This is to allow workspace users to get an overview of all
changes of all mounts by navigating to the virtual page uid=0.

Releases: master, 11.5, 10.4
Resolves: #95854
Related: #95972
Related: #91145
Change-Id: I7f6370f327711396193cf56b63f15876350c2559
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72028
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>