[BUGFIX] Save correct table name as ref_table in refindex
The identifier from the linkHandler page TSconfig was formerly used as
the reference table name "ref_table". This identifier is free to choose
and shouldn't be interpreted as a name of a table.

In order to resolve the correct table name, which is stated in the
configuration, the page id of the reference record is now used. This
allows the TypolinkSoftReferenceParser to read the according page
TSconfig and to set the correct table name.

Resolves: #95380
Releases: master, 10.4
Change-Id: I28817a7267a107df27828d99f7e0cf2eeaaf2de1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72202
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Henrik Ziegenhain <henrik@ziegenhain.me>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Henrik Ziegenhain <henrik@ziegenhain.me>
Reviewed-by: Benni Mack <benni@typo3.org>