[TASK] runTests.sh: xdebug trigger works with macOS and WSL2
docker on mac and windows WSL2 encapsulate a VM into
the system. The current xdebug remote config does not
work in those cases since the host IP can not be
determined with the current solution.

Switching to host.docker.internal works for macOS
and WSL2, and works for linux too in combination with
an extra_hosts setting.

This requires linux to use at least docker 20.10,
which shouldn't be a huge issue with casual developer
machines.

Resolves: #95999
Resolves: #94932
Releases: master, 10.4
Change-Id: I02acfeefc08f338841b4fecdb952c7da37ff65e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72194
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>