[FEATURE] Allow tagging of data processors
This adds the possibility to tag data processors
with the "data.processor" tag in the service configuration.

Tagged services will automatically be added to the new
DataProcessorRegistry, which enables DI out of the box.
As a side effect, this allows to define an "identifier",
which can be used as an alias in e.g. TypoScript
configurations.

Note: Tagged services are not automatically set public.
This only happens implicit when fetched via the registry.
In case one still uses the FQCN instead of the alias /
identifier, the service must still be set "public: true"
manually.

All core data processors are tagged and defined by an
identifier. A possible migration in TypoScript would
therefore look like:

Default:
10 = TYPO3\CMS\Frontend\DataProcessing\DatabaseQueryProcessor

New possibility:
10 = database-query

Resolves: #96005
Related: #92910
Releases: main
Change-Id: I678b899c9a8f152b1fafc71fb0abb0eadfebb885
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72199
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>