[BUGFIX] Add missing checks for translate keys in compare view
The compare view of EXT:beuser does a couple
of f:translate calls with dynamic values as key.
Since some of those calls did not properly check
the provided key, before passing it to the view
helper, this previously led to exceptions.

This is now fixed by properly checking the
key, before passing it to the f:translate
view helper. Additionally, some defaults
are added to support static labels.

Resolves: #94202
Releases: master, 11.5
Change-Id: Iff14f7c3b17108288227b108b5005d28611e20a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72248
Tested-by: core-ci <typo3@b13.com>
Tested-by: David Menzel <foxxdav@yahoo.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: David Menzel <foxxdav@yahoo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>