[BUGFIX] Do not use hardcoded field name for belayoutwizard
TCA renderType=belayoutwizard did previously not
work for custom fields, as can be seen in styleguide
`tx_styleguide_elements_basic.text_20`.

BackendLayoutWizardElement used the hardcoded
field name "config", which however is only used for
the "backend_layout" table. Additionally, using the
raw database value instead of the processed
"itemFormElValue" led to further issues, e.g. when
the field is defined with "l10n_display=defaultAsReadonly".

This is now fixed by using the "itemFormElValue",
since this is the correct value, which is also added
to the corresponding input field and would also
contain the "default" for new records.

Additionally, the given parsed itemFormElValue is
now checked to start with the correct root level
key "backend_layout.". This prevents undefined
array key warnings, triggered when a malformed
default value is used for the field.

Resolves: #96060
Releases: master, 11.5
Change-Id: I2d7c40bf3b33d6d2528c7575febe67d105ab5e12
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72266
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>