[BUGFIX] Observe GridEditor visibility to refresh codeMirror
In case a TCA field with renderType=belayoutwizard
is configured to be not in the initially visible tab, the
CodeMirror editor remains empty, when switching to
the tab, containing the editor. The reason is, CodeMirror
needs to be refreshed, after becoming visible.

Therefore, the GridEditor, holding the configuration,
is now observed, if not initially visible. As soon as it
becomes visible, the CodeMirror instance - if any -
is refreshed.

Can be tested with `tx_styleguide_elements_basic.text_20`.

Resolves: #96062
Releases: master, 11.5
Change-Id: Ic03c315e8f12bc2f3623fdb11a1782d5c1081a23
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72277
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>