[BUGFIX] Improve legacy handling of not Composer installed extensions
Starting with 11.4, in Composer managed TYPO3 installations,
having extensions, that are not installed by Composer, are deprecated.

They are still detected and registered though, by scanning the
typo3conf/ext folder and registering non Composer installed extensions
from there.

This process is now optimised to only read composer.json files
of extensions, that are known to not be installed via Composer,
instead of scanning all of them.

In that context it is now also ensured, that the root package,
in case it being an extension, is treated as "Composer installed".

Releases: main, 11.5
Resolves: #96076
Change-Id: I6544d43bf8ce2d49d5d480aab8baba4cdf23347e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72426
Tested-by: core-ci <typo3@b13.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Helmut Hummel <typo3@helhum.io>