[BUGFIX] Update CSV references when discarding record
When an element is discarded in workspaces that is referenced
by another record in a group CSV field, those referencing
records need to be updated to no longer list the discarded
record.

The patch adds a method to DataHandler discard logic to
take care of that.

This resolves the very last call setting assertCleanReferenceIndex
to false. That test property has been a temporary solution and
is dropped now, tests extending AbstractDataHandlerActionTestCase
can no longer (easily) prevent the tearDown() refindex check and
should never do that again.

Change-Id: Ifd2e5893d2d39b77762920b8b41ab547ed20ecd7
Resolves: #96080
Releases: master, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72297
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>