[BUGFIX] Fix includeCSS.inline for files with EXT: syntax
The feature page.includeCSS.inline checks the existence
of the given file with `is_file()`. This does not work
for files included with the EXT: prefix, since the
path to the filename is not fully resolved any more
since https://review.typo3.org/c/Packages/TYPO3.CMS/+/71457

This patch processes the given file using the
`getStreamlinedFileName` function in order to resolve
the relative path to the file.

Resolves: #96098
Releases: master, 11.5
Change-Id: I2958eb13228622e724b5d4db1919a982fcd6247e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72308
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>