[BUGFIX] Create auto-redirects with default language page uid
Creating auto redirects, if page slug is changed, has been extended
to add the target with LinkHandler 't3://' uri schema with #89327,
adding language argument '_language' to enforce the language for the
redirect, missing that that did not worked, which was solved through
adding '_language' argument handling in #96043.

As languge arguments '_language' and 'L' are ignored, if the provided
page uid is already the uid of the localized page, the parameter was
meaning less. Further this breaks language fallback handling, if the
localized page record is deleted.

This patch uses now the default language page uid for creating target
of auto-redirects to mitigate the issue that redirect will break if
localized page is deleted but language fallback is configured.

Resolves: #96103
Related: #96043
Related: #89327
Releases: master, 11.5, 10.4
Change-Id: I2ece06bf053bda4edb542910c3f51b1e6c8e052c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72327
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>