[BUGFIX] UploadedResourceViewHelper does not extend UploadViewHelper
With #95486, the UploadViewHelper got extended
for a new tag attribute "accept", which can be used
to limit the allowed file types for an upload field. The
EXT:form UploadedResourceViewHelper, which extends
the UploadViewHelper, registers the "accept" argument
and if given, implodes the value and adds it as attribute
to the tag. This however now led to an exception since
the "accept" attribute got already added by the parent
class.

This is fixed by no longer extending UploadViewHelper,
but implementing the required logic directly in the
UploadedResourceViewHelper.

This step would have been necessary in v12 anyways,
see #95298.

Resolves: #96146
Related: #95298
Related: #95486
Releases: main
Change-Id: Idbc1e284d9a7097fd7397562e566d32debc5f030
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72383
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>