[FEATURE] Show registered toolbar items in configuration module
With #96041 the registration of toolbar items was
changed. All implementations of the corresponding
interface are now automatically registered. The
$GLOBALS configuration does no longer have
any effect. An administrator however was no
longer able to get an overview of the registered
toolbar items in the configuration module.

Therefore, this patch improves the situation
by adding a new configuration provider, which
in contrast to the GlobalVariableProvider,
already shows the final ordering.

Resolves: #96152
Related: #96041
Releases: main
Change-Id: Ieba1988e65fa777aa0c71c258c0bac9e76a10ecb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72392
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>