[TASK] gitlab-ci: Out-of-order nightly jobs
Usually, when running jobs in gitlab-ci, they're
organized in stages - jobs of a second stage only
execute when all jobs of the previous stage were
successful.

The pre-merge jobs use this to supress jobs in case
the "early" stage checking cgl and composer
fail. This helps suppressing waste of resources.

The nightly jobs are organized in more logical stages.
The stage supression isn't too useful here. Pipelines
are only executed as scheduled pipelines or manually
to sometimes test if the extended nightly jobs are
happy with a single patch before merging it.

To reduce runtime of nightlies, they now introduce
a gitlab-ci feature for single stage jobs that ignore
stage boundaries. With this out-of-order configuration
the stage separation is basically ignored, which on one
side reduces overall runtime of nightlies since more
jobs are executed in parallel and on the other side now
find all issues in case multiple merged patched kill a
nigthly and one hides fails from the other away due to
stage constraints.

Change-Id: I8e461592eecd0a66b0fcf6ee5ab14509ebec8277
Resolves: #96166
Releases: master, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72409
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>