[FEATURE] Introduce RedirectWasHitEvent
A new PSR-14 based event "RedirectWasHitEvent" is introduced,
allowing extension authors to further process a matched redirect
and to adjust the PSR-7 Response. It's furthermore possibility
to influence core functionality, e.g. the hit count increment.
This is especially useful for e.g. the use of custom monitoring
tools, which should not increment the hit count while testing
the redirects.

As a side effect, the middleware is cleaned up, since the
increment task is moved into a dedicated event listener.

Resolves: #96147
Releases: main
Change-Id: I146b60d9a347199f7a871edca0845e7d635817df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72437
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>