[BUGFIX] Move extension scanner entry of arrayDiffAssocRecursive
With #94137, the ArrayUtility::arrayDiffAssocRecursive
method was adjusted to actually behave like `array_diff_assoc`.

To not introduce a breaking change after 11.0, a third
parameter was added to enable the correct behaviour.
This parameter however will be removed again in v12.

Therefore, the extension scanner entry hast to be
placed in MethodArgumentDroppedStaticMatcher
instead of MethodCallStaticMatcher, because only
the last parameter is deprecated, not the whole
method.

Resolves: #96193
Related: #94137
Releases: main, 11.5
Change-Id: Ic6844b509046037f411b302eff2ee983bf093917
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72453
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>