[BUGFIX] Do not add error in TypoScriptParser if modifier returns null
Currently, the TypoScriptParser adds an error when
the result of a modifier is null. This is, because
the line is not set to a new value in such a case
and the following switch statement is not expecting
a line with a value modifier.

This has been fixed by skipping further checks on
the value, if the modifier returns null.

Resolves: #94109
Releases: master, 11.5
Change-Id: I28ae1570a5b0b4fc5ad667bacd86a17234aa1c60
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72468
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>