[BUGFIX] Check extension version number format
A false check in the TER API allowed to use
arbitrary version strings, as soon as they start
with the known `XXX.XXX.XXX` schema. This
will be fixed in TER API. However, core as receiver
should also validate the incoming data, before
storing them in the database.

Therefore, the same version number check,
which is used in the TER API, is now also
applied in the ExtensionXmlParser.

Resolves: #96209
Releases: main, 11.5, 10.4
Change-Id: I487eed34c6b25a5e3841d5263b52b9a73aaf681a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72477
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>