[!!!][TASK] Remove sys_language database table
The database table sys_language has become obsolete
with the introduction of site handling and site
languages.

Therefore, the sys_language database table and the
corresponding TCA is now removed. Fixtures and tests
are adjusted to no longer rely on this table.

Additionally, also a couple of code comments through
the Core are adjusted.

Executed commands:

   composer u typo3/cms-styleguide

Resolves: #96277
Related: #94165
Releases: main
Change-Id: I4ab0f3c7bb6de110cc1ce9826b2718f60d8a18a0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72534
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>