[BUGFIX] Exclude colorpicker from overflow hidden
In order to fix a visual bug of overflowing select
fields inside palettes, the css attribute overflow:
hidden has been added to one of the wrapping fields
of form wizards.

This caused our colorpicker to be invisible for the
user, as it is directly appended after the input
field inside the DOM (Unlike to the behaviour of
flatpickr, which is added to the end of the body
tag).

A new modifier class
"form-wizards-element--overflow-visible" has been
added to disable the overflow hidden rule for
specific fields.

Resolves: #96130
Related: #94688
Releases: main, 11.5
Change-Id: I93385c153abc2ff105e1e8c432be9aea3a2be576
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72544
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>