[TASK] Allow composer installers ^4.0
Allow composer installers ^4.0 and also enforce some constraints
that could be violated with the new installers.

1. Do not allow duplicate extension keys

Multiple packages could define the same extension key.
When all extensions are then installed in vendor folder, there won't
be any conflicts on the file system, so an exception is now thrown
to make people aware of this misconfiguration.

2. Throw an exception if extensions of type typo3-cms-extension do
   not define an extension key

Actually this exception is thrown in the installers package, but the
artifact builder must propagate this exception to ensure it
reaches the users.

composer req typo3/cms-composer-installers:"^2.0 || ^3.0 || ^4.0"
composer req typo3/cms-composer-installers:"^2.0 || ^3.0 || ^4.0" \
 --no-update -d typo3/sysext/core/

Resolves: #96276
Releases: main, 11.5
Change-Id: Ib76e829f09dbf095fb01031389e301c6e7742103
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72546
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Helmut Hummel <typo3@helhum.io>