[BUGFIX] Respect single tags without white space
Make HTML parser more tolerant for single tags with ending slash but
without white spaces. So `<hr/>` already was allowed and internally
converted to `<hr />` while parsing the HTML. This patch extends this
replace rule to all tags so that also `<br/>` can be used in RTE fields.

Resolves: #96307
Releases: main, 11.5, 10.4
Change-Id: I7ba6c88e35804c6b5f991fb6b730fbc3d7a2514e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72597
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>